remove mixin usage, add fortawesome/ember-font-awesome, remove tooltip dep
