+++ .editorconfig
@@ -10,7 +10,7 @@ charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
 indent_style = space
-indent_size = 2
+indent_size = 4
 
 [*.hbs]
 insert_final_newline = false
