+++ addon/templates/components/button-basic.hbs
@@ -1,22 +1,9 @@
 {{#if hasIcon}}
-    {{!-- we need to explicity reference this.propName here --}}
-    {{!-- to avoid a naming collision with a 3rd party addon --}}
-    {{!-- that has a helper named "icon" --}}
-    {{fa-icon this.icon class="action-button-icon"}}
+    <FaIcon @icon={{this.icon}} class="action-button-icon" />
 {{/if}}
-{{#if (not iconBtn)}}
-    {{#if label}}
-        <span class="action-button-lbl">{{label}}</span>
+{{#if (not this.iconBtn)}}
+    {{#if @label}}
+        <span class="action-button-lbl">{{@label}}</span>
     {{/if}}
     {{yield}}
-{{/if}}
-{{#if tooltip}}
-    {{#attach-tooltip
-        animation=tooltip-animation
-        arrow=tooltip-arrow
-        hideOn=tooltip-hide-on
-        registerAPI=(action "onTooltipApiChange")
-    }}
-        {{tooltip}}
-    {{/attach-tooltip}}
 {{/if}}
