+++ package.json
@@ -25,15 +25,15 @@
     "postpublish": "ember ts:clean"
   },
   "dependencies": {
-    "ember-attacher": "^0.13.5",
+    "@fortawesome/ember-fontawesome": "^0.2.1",
     "ember-cli-babel": "^7.17.2",
     "ember-cli-htmlbars": "^4.2.2",
     "ember-cli-typescript": "^3.1.3",
-    "ember-font-awesome": "4.0.0-rc.3",
     "ember-truth-helpers": "^2.0.0"
   },
   "devDependencies": {
     "@ember/optional-features": "^1.3.0",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
     "@glimmer/component": "^1.0.0",
     "@glimmer/tracking": "^1.0.0",
     "@types/ember": "^3.1.1",
