+++ tests/dummy/app/templates/application.hbs
@@ -1,34 +1,33 @@
 <div class="p-3">
     <h2 id="title">Welcome to Ember</h2>
-    {{button-basic
-        type="primary"
-        icon="remove"
-        label="Remove"
-        tooltip="Hi! I'm a tooltip!"
-        autofocus=true
-        tabindex=2
-    }}
-    {{button-basic
-        type="success"
-        outline=true
-        icon="save"
-        label="Save"
-        autofocus=false
-        tabindex=3
-    }}
-    {{button-basic
-        type="danger"
-        icon="ban"
-        label="Cancel"
-        autofocus=false
-        tabindex=2
-    }}
-    {{button-basic
-        type="danger"
-        outline=true
-        icon="ban"
-        label="Untabable"
-        tabindex=-1
-    }}
+    <ButtonBasic
+        @type="primary"
+        @icon="user"
+        @label="Remove"
+        @autofocus={{true}}
+        @tabindex={{2}}
+    />
+    <ButtonBasic
+        @type="success"
+        @outline={{true}}
+        @icon="save"
+        @label="Save"
+        @autofocus={{false}}
+        @tabindex={{3}}
+    />
+    <ButtonBasic
+        @type="danger"
+        @icon="ban"
+        @label="Cancel"
+        @autofocus={{false}}
+        @tabindex={{2}}
+    />
+    <ButtonBasic
+        @type="danger"
+        @outline={{true}}
+        @icon="ban"
+        @label="Un-tabbable"
+        @tabindex={{-1}}
+    />
     {{outlet}}
 </div>
